sentinel-ai/requirements.txt
pandas
numpy
scipy
scikit-learn
psycopg2-binary
python-dotenv
